warnings fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@16933 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/DbEditorPP/src/main_window.cpp
@@ -274,7 +274,7 @@ INT_PTR CALLBACK MainDlgProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam)
 			if ((HWND)lParam == GetDlgItem(hwnd, IDC_SPLITTER)) {
 				RECT rc;
 				GetClientRect(hwnd, &rc);
-				POINT pt = { wParam, 0 };
+				POINT pt = { (LONG)wParam, 0 };
 				ScreenToClient(hwnd, &pt);
 
 				splitterPos = rc.left + pt.x + 1;
